Deal with slashes in object keys and add tests
Removes the assumption that the object key won't consist of the '/'
character (which is both perfectly acceptable and a common method of
organisation). Since we previously had the version as the last part of
the path, we now move that to a query parameter.

Also adds some automated test cases so we can catch bugs early.