Increase default blocksize to 256 kB
The basesrc default is 4 kB, which is really small and makes reads very
inefficient (every block translates to an HTTP GET request). We bump
this up to 256 kB, which should be better.

This is still somewhat arbitrary -- depending on the usage, it could
likely be higher still, but we pick a somewhat conservative number for
now.